+++ contracts/asset-proxy/src/artifacts.ts
@@ -9,6 +9,7 @@ import * as BalancerBridge from '../generated-artifacts/BalancerBridge.json';
 import * as BancorBridge from '../generated-artifacts/BancorBridge.json';
 import * as ChaiBridge from '../generated-artifacts/ChaiBridge.json';
 import * as CreamBridge from '../generated-artifacts/CreamBridge.json';
+import * as CryptoComBridge from '../generated-artifacts/CryptoComBridge.json';
 import * as CurveBridge from '../generated-artifacts/CurveBridge.json';
 import * as DexForwarderBridge from '../generated-artifacts/DexForwarderBridge.json';
 import * as DODOBridge from '../generated-artifacts/DODOBridge.json';
@@ -77,6 +78,7 @@ export const artifacts = {
     BancorBridge: BancorBridge as ContractArtifact,
     ChaiBridge: ChaiBridge as ContractArtifact,
     CreamBridge: CreamBridge as ContractArtifact,
+    CryptoComBridge: CryptoComBridge as ContractArtifact,
     CurveBridge: CurveBridge as ContractArtifact,
     DODOBridge: DODOBridge as ContractArtifact,
     DexForwarderBridge: DexForwarderBridge as ContractArtifact,