+++ contracts/zero-ex/contracts/src/transformers/bridges/mixins/MixinAdapterAddresses.sol
@@ -26,6 +26,7 @@ contract MixinAdapterAddresses
         address balancerBridge;
         address creamBridge;
         address curveBridge;
+        address cryptoComBridge;
         address dodoBridge;
         address kyberBridge;
         address mooniswapBridge;