+++ packages/asset-swapper/src/constants.ts
@@ -75,6 +75,7 @@ const DEFAULT_EXCHANGE_PROXY_EXTENSION_CONTRACT_OPTS: ExchangeProxyContractOpts 
     },
     refundReceiver: NULL_ADDRESS,
     isMetaTransaction: false,
+    shouldSellEntireBalance: false,
 };
 
 const DEFAULT_FORWARDER_SWAP_QUOTE_EXECUTE_OPTS: SwapQuoteExecutionOpts = DEFAULT_FORWARDER_SWAP_QUOTE_GET_OPTS;