+++ packages/asset-swapper/src/quote_consumers/exchange_proxy_swap_quote_consumer.ts
@@ -34,7 +34,7 @@ import { getSwapMinBuyAmount } from './utils';
 
 // tslint:disable-next-line:custom-no-magic-numbers
 const MAX_UINT256 = new BigNumber(2).pow(256).minus(1);
-const { NULL_ADDRESS, ZERO_AMOUNT } = constants;
+const { NULL_ADDRESS, NULL_BYTES, ZERO_AMOUNT } = constants;
 
 export class ExchangeProxySwapQuoteConsumer implements SwapQuoteConsumerBase {
     public readonly provider: ZeroExProvider;
@@ -90,7 +90,7 @@ export class ExchangeProxySwapQuoteConsumer implements SwapQuoteConsumerBase {
             ...opts.extensionContractOpts,
         };
         // tslint:disable-next-line:no-object-literal-type-assertion
-        const { refundReceiver, affiliateFee, isFromETH, isToETH } = optsWithDefaults;
+        const { refundReceiver, affiliateFee, isFromETH, isToETH, shouldSellEntireBalance } = optsWithDefaults;
 
         const sellToken = getTokenFromAssetData(quote.takerAssetData);
         const buyToken = getTokenFromAssetData(quote.makerAssetData);
@@ -131,6 +131,26 @@ export class ExchangeProxySwapQuoteConsumer implements SwapQuoteConsumerBase {
             };
         }
 
+        if (isDirectSwapCompatible(quote, optsWithDefaults, [ERC20BridgeSource.LiquidityProvider])) {
+            const target = quote.orders[0].makerAddress;
+            return {
+                calldataHexString: this._exchangeProxy
+                    .sellToLiquidityProvider(
+                        isFromETH ? ETH_TOKEN_ADDRESS : sellToken,
+                        isToETH ? ETH_TOKEN_ADDRESS : buyToken,
+                        target,
+                        NULL_ADDRESS,
+                        sellAmount,
+                        minBuyAmount,
+                        NULL_BYTES,
+                    )
+                    .getABIEncodedTransactionData(),
+                ethAmount: isFromETH ? sellAmount : ZERO_AMOUNT,
+                toAddress: this._exchangeProxy.address,
+                allowanceTarget: this.contractAddresses.exchangeProxyAllowanceTarget,
+            };
+        }
+
         // Build up the transforms.
         const transforms = [];
         if (isFromETH) {
@@ -139,7 +159,7 @@ export class ExchangeProxySwapQuoteConsumer implements SwapQuoteConsumerBase {
                 deploymentNonce: this.transformerNonces.wethTransformer,
                 data: encodeWethTransformerData({
                     token: ETH_TOKEN_ADDRESS,
-                    amount: sellAmount,
+                    amount: shouldSellEntireBalance ? MAX_UINT256 : sellAmount,
                 }),
             });
         }
@@ -155,7 +175,7 @@ export class ExchangeProxySwapQuoteConsumer implements SwapQuoteConsumerBase {
                     buyToken: intermediateToken,
                     side: FillQuoteTransformerSide.Sell,
                     refundReceiver: refundReceiver || NULL_ADDRESS,
-                    fillAmount: firstHopOrder.takerAssetAmount,
+                    fillAmount: shouldSellEntireBalance ? MAX_UINT256 : firstHopOrder.takerAssetAmount,
                     maxOrderFillAmounts: [],
                     rfqtTakerAddress: NULL_ADDRESS,
                     orders: [firstHopOrder],
@@ -177,6 +197,7 @@ export class ExchangeProxySwapQuoteConsumer implements SwapQuoteConsumerBase {
                 }),
             });
         } else {
+            const fillAmount = isBuyQuote(quote) ? quote.makerAssetFillAmount : quote.takerAssetFillAmount;
             transforms.push({
                 deploymentNonce: this.transformerNonces.fillQuoteTransformer,
                 data: encodeFillQuoteTransformerData({
@@ -184,7 +205,7 @@ export class ExchangeProxySwapQuoteConsumer implements SwapQuoteConsumerBase {
                     buyToken,
                     refundReceiver: refundReceiver || NULL_ADDRESS,
                     side: isBuyQuote(quote) ? FillQuoteTransformerSide.Buy : FillQuoteTransformerSide.Sell,
-                    fillAmount: isBuyQuote(quote) ? quote.makerAssetFillAmount : quote.takerAssetFillAmount,
+                    fillAmount: shouldSellEntireBalance ? MAX_UINT256 : fillAmount,
                     maxOrderFillAmounts: [],
                     rfqtTakerAddress: NULL_ADDRESS,
                     orders: quote.orders,
@@ -238,7 +259,7 @@ export class ExchangeProxySwapQuoteConsumer implements SwapQuoteConsumerBase {
             .transformERC20(
                 isFromETH ? ETH_TOKEN_ADDRESS : sellToken,
                 isToETH ? ETH_TOKEN_ADDRESS : buyToken,
-                sellAmount,
+                shouldSellEntireBalance ? MAX_UINT256 : sellAmount,
                 minBuyAmount,
                 transforms,
             )
@@ -289,6 +310,10 @@ function isDirectSwapCompatible(
     }
     const fill = order.fills[0];
     if (!directSources.includes(fill.source)) {
+        return false;
+    }
+    // VIP does not support selling the entire balance
+    if (opts.shouldSellEntireBalance) {
         return false;
     }
     return true;