+++ contracts/erc721/CHANGELOG.md
@@ -5,6 +5,10 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v3.1.16 - _December 3, 2020_
+
+    * Dependencies updated
+
 ## v3.1.15 - _November 19, 2020_
 
     * Dependencies updated